fix(webapp): hopefully fixing auth forever
moving away from bcrypt; using crypto instead

+++ packages/webapp/src/pages/api/auth/logout.ts
@@ -1,4 +1,3 @@
-import bcrypt from 'bcryptjs';
 import type { APIRoute } from 'astro';
 
 import { prisma } from '@utils/database';
@@ -9,17 +8,20 @@ export const get: APIRoute = async ({ params, request, cookies }) => {
 
   // if session cookie is missing, return error
   if (!sessionId) {
-    return new Response(JSON.stringify({message: 'not logged in'}), {status: 401});
+    return new Response(JSON.stringify({ message: 'not logged in' }), { status: 401 });
   }
 
   // delete the session from the database
-  await prisma.session.delete({
-    where: { id: sessionId },
-  });
+  const sessionExists = await prisma.session.findFirst({ where: { id: sessionId } });
+  if (sessionExists) {
+    await prisma.session.delete({
+      where: { id: sessionId },
+    });
+  }
 
   // delete the session cookie
-  cookies.delete('session', {path: '/'});
+  cookies.delete('session', { path: '/' });
 
   // return a success error
-  return new Response(JSON.stringify({message: 'logged out'}), {status: 200});
+  return new Response(JSON.stringify({ message: 'logged out' }), { status: 200 });
 }
